pycassandra/requirements.txt
Flask==0.11.1
cassandra-driver
